+++ resources/views/components/table/datarow.blade.php
@@ -90,6 +90,11 @@
         @elseif ($key == 'status')
             {{-- true/false/null als Haken, Kreuz oder Strich. --}}
             <td scope="row" class="py-2.5 px-4"><x-statusicon :value="$value" /></td>
+        @elseif ($key == 'einladung')
+            {{-- $value ist der Benutzer selbst: Ob die Einladung offen oder
+                 abgelaufen ist, haengt an der Frist aus config/auth.php - das
+                 weiss das Modell, nicht die Tabelle. --}}
+            <td scope="row" class="py-2.5 px-4 text-gray-900 dark:text-gray-100"><x-einladung :user="$value" /></td>
         @elseif ($key == 'zweitestufe')
             {{-- Eigener Schluessel und nicht 'status': Dort gibt es nur an, aus
                  und unbekannt - "verlangt, aber noch nicht eingerichtet" ist